// src/app/models/hackathons.ts
export interface Hackathons {
    participant_name: string;
    id_number: number;
    email_address: string;
    whatsapp_number: number;
    disability: string;
    disability_description: string;
    update_participat: string;
    development_experience: string;
    familiar_language: string;
    programming_experience: string;
    cloudprovider_expertise: string;
    hackathons_participation: string;
    which_hackathons: string;
    sourceofinformation: string;
}